[fix]: fixed KYC page to send request when loaded
then wait for response. once response is obtained, stop the loading screen and show the button to proceed.

Note that, the problem of sending two requests to kycValidate when /kyc is loaded still exists. this behaviour can be observed by looking at the server's output or on browser's dev page. giving a hint that the /kyc page might have been loaded twice in a short time, possibly a problem with the service